+++ app/g/[id]/GameClient.tsx
@@ -3,12 +3,10 @@
 import { useCallback, useEffect, useState } from "react";
 import { useLocalGroups } from "@/hooks/useLocalGroups";
 import { roll } from "@/lib/algorithm";
-import Header from "@/components/Header";
-import PlayerCard from "@/components/PlayerCard";
-import AddPlayerPicker from "@/components/AddPlayerPicker";
-import RollingCard, { RollState } from "@/components/RollingCard";
-import Confetti from "@/components/Confetti";
 import { GamePhase, RosterPlayer, SessionPlayer } from "@/types/game";
+import SetupPhase from "./SetupPhase";
+import RollingPhase from "./RollingPhase";
+import CompletePhase from "./CompletePhase";
 
 interface Props {
   group: { id: string; name: string };
@@ -21,18 +19,19 @@ export default function GameClient({ group, initialRoster }: Props) {
   const [phase, setPhase] = useState<GamePhase>("setup");
   const [roster, setRoster] = useState<RosterPlayer[]>(initialRoster);
   const [sessionPlayers, setSessionPlayers] = useState<SessionPlayer[]>([]);
-  const [showPicker, setShowPicker] = useState(false);
 
   useEffect(() => {
     addGroup(group.id, group.name);
   }, [group.id, group.name, addGroup]);
 
-  // ── Setup helpers ─────────────────────────────────────────────────────────
+  // ── Derived ───────────────────────────────────────────────────────────────
 
   const prices = sessionPlayers.map((p) => parseFloat(p.price) || 0);
   const total = prices.reduce((a, b) => a + b, 0);
   const probabilities = prices.map((p) => (total > 0 ? p / total : 0));
   const canStart = prices.filter((p) => p > 0).length >= 2;
+
+  // ── Setup callbacks ───────────────────────────────────────────────────────
 
   function updatePrice(playerId: string, value: string) {
     setSessionPlayers((prev) =>
@@ -49,7 +48,6 @@ export default function GameClient({ group, initialRoster }: Props) {
       ...prev,
       { playerId: player.id, name: player.name, price },
     ]);
-    setShowPicker(false);
   }
 
   async function addNewToSession(name: string, price: string) {
@@ -66,10 +64,9 @@ export default function GameClient({ group, initialRoster }: Props) {
       ...prev,
       { playerId: newPlayer.id, name: newPlayer.name, price },
     ]);
-    setShowPicker(false);
   }
 
-  // ── Rolling helpers ───────────────────────────────────────────────────────
+  // ── Rolling callbacks ─────────────────────────────────────────────────────
 
   function rollPlayer(playerId: string) {
     const player = sessionPlayers.find((p) => p.playerId === playerId)!;
@@ -116,189 +113,46 @@ export default function GameClient({ group, initialRoster }: Props) {
     });
   }
 
-  function getRollState(player: SessionPlayer): RollState {
-    if (player.isRolling) return "rolling";
-    if (player.rolledScore === undefined) return "waiting";
-    // Among players who have rolled, find the minimum score
-    const rolledScores = sessionPlayers
-      .filter((p) => p.rolledScore !== undefined && !p.isRolling)
-      .map((p) => p.rolledScore!);
-    const minScore = Math.min(...rolledScores);
-    return player.rolledScore === minScore ? "danger" : "safe";
-  }
-
-  // ── Setup phase ───────────────────────────────────────────────────────────
+  // ── Render ────────────────────────────────────────────────────────────────
 
   if (phase === "setup") {
     return (
-      <div className="min-h-screen flex flex-col">
-        <Header groupId={group.id} groupName={group.name} />
-
-        <main className="flex-1 flex flex-col max-w-2xl mx-auto w-full">
-          {/* Scrollable player area */}
-          <div className="flex-1 overflow-y-auto p-4 pb-2 flex flex-col gap-6">
-            <h2 className="font-display text-gold/70 tracking-widest text-xs text-center pt-2">
-              TODAY'S LUNCH
-            </h2>
-
-            {sessionPlayers.length > 0 && (
-              <div className="grid grid-cols-2 gap-3">
-                {sessionPlayers.map((player, i) => (
-                  <PlayerCard
-                    key={player.playerId}
-                    name={player.name}
-                    price={player.price}
-                    probability={probabilities[i]}
-                    onPriceChange={(v) => updatePrice(player.playerId, v)}
-                    onRemove={() => removeFromSession(player.playerId)}
-                  />
-                ))}
-              </div>
-            )}
-
-            {sessionPlayers.length === 0 && (
-              <div className="flex-1 flex flex-col items-center justify-center gap-2 py-16">
-                <p className="text-4xl opacity-20">🎲</p>
-                <p className="text-cream/30 text-sm">No players yet</p>
-              </div>
-            )}
-          </div>
-
-          {/* Sticky bottom bar */}
-          <div className="p-4 pt-2 flex flex-col gap-3 border-t border-gold/10">
-            {total > 0 && (
-              <div className="text-center">
-                <p className="text-cream/30 text-xs uppercase tracking-widest">Total bill</p>
-                <p className="text-gold font-display text-2xl">${total.toFixed(2)}</p>
-              </div>
-            )}
-            <button
-              onClick={() => setShowPicker(true)}
-              className="w-full border border-dashed border-gold/20 text-cream/40
-                         hover:border-gold/40 hover:text-cream/70
-                         rounded-lg py-3 text-sm transition-colors"
-            >
-              + Add player
-            </button>
-
-            <button
-              onClick={() => setPhase("rolling")}
-              disabled={!canStart}
-              className="w-full bg-gold text-casino-black font-display tracking-widest
-                         text-sm py-4 rounded-lg transition-colors
-                         hover:bg-gold-light
-                         disabled:opacity-30 disabled:cursor-not-allowed"
-            >
-              START ROLLING 🎲
-            </button>
-          </div>
-        </main>
-
-        {showPicker && (
-          <AddPlayerPicker
-            roster={roster}
-            sessionPlayerIds={sessionPlayers.map((p) => p.playerId)}
-            onAddExisting={addExistingToSession}
-            onAddNew={addNewToSession}
-            onClose={() => setShowPicker(false)}
-          />
-        )}
-      </div>
+      <SetupPhase
+        group={group}
+        sessionPlayers={sessionPlayers}
+        roster={roster}
+        probabilities={probabilities}
+        total={total}
+        canStart={canStart}
+        onUpdatePrice={updatePrice}
+        onRemove={removeFromSession}
+        onAddExisting={addExistingToSession}
+        onAddNew={addNewToSession}
+        onStart={() => setPhase("rolling")}
+      />
     );
   }
-
-  // ── Rolling phase ─────────────────────────────────────────────────────────
 
   if (phase === "rolling") {
     return (
-      <div className="min-h-screen flex flex-col">
-        <Header groupId={group.id} groupName={group.name} />
-
-        <main className="flex-1 flex flex-col max-w-2xl mx-auto w-full">
-          <div className="flex-1 overflow-y-auto p-4 pb-2 flex flex-col gap-6">
-            <h2 className="font-display text-gold/70 tracking-widest text-xs text-center pt-2">
-              ROLL YOUR DICE
-            </h2>
-
-            <div className="grid grid-cols-2 gap-3">
-              {sessionPlayers.map((player) => (
-                <RollingCard
-                  key={player.playerId}
-                  player={player}
-                  rollState={getRollState(player)}
-                  onRoll={() => rollPlayer(player.playerId)}
-                  onDieComplete={() => onDieComplete(player.playerId)}
-                />
-              ))}
-            </div>
-          </div>
-
-          <div className="p-4 pt-2 border-t border-gold/10 text-center">
-            <p className="text-cream/30 text-xs uppercase tracking-widest">Total bill</p>
-            <p className="text-gold font-display text-2xl">${total.toFixed(2)}</p>
-          </div>
-        </main>
-      </div>
+      <RollingPhase
+        group={group}
+        sessionPlayers={sessionPlayers}
+        total={total}
+        onRoll={rollPlayer}
+        onDieComplete={onDieComplete}
+      />
     );
   }
 
-  // ── Complete phase ────────────────────────────────────────────────────────
-
-  const payer = sessionPlayers.reduce((min, p) =>
-    (p.rolledScore ?? Infinity) < (min.rolledScore ?? Infinity) ? p : min
-  );
-  const billTotal = sessionPlayers.reduce((sum, p) => sum + parseFloat(p.price), 0);
-
-  function newGame() {
-    setPhase("setup");
-    setSessionPlayers([]);
-  }
-
-  function getCompleteState(player: SessionPlayer): RollState {
-    return player.playerId === payer.playerId ? "payer" : "safe";
-  }
-
   return (
-    <div className="min-h-screen flex flex-col">
-      <Confetti />
-      <Header groupId={group.id} groupName={group.name} />
-
-      <main className="flex-1 flex flex-col max-w-2xl mx-auto w-full">
-        <div className="flex-1 overflow-y-auto p-4 pb-2 flex flex-col gap-6">
-          <h2 className="font-display text-gold/70 tracking-widest text-xs text-center pt-2">
-            RESULT
-          </h2>
-
-          <div className="grid grid-cols-2 gap-3">
-            {sessionPlayers.map((player) => (
-              <RollingCard
-                key={player.playerId}
-                player={player}
-                rollState={getCompleteState(player)}
-                onRoll={() => {}}
-                onDieComplete={() => {}}
-              />
-            ))}
-          </div>
-
-          {/* Total bill */}
-          <div className="text-center">
-            <p className="text-cream/30 text-xs uppercase tracking-widest">Total bill</p>
-            <p className="text-gold font-display text-2xl">${billTotal.toFixed(2)}</p>
-          </div>
-        </div>
-
-        {/* Bottom bar */}
-        <div className="p-4 pt-2 border-t border-gold/10">
-          <button
-            onClick={newGame}
-            className="w-full bg-gold text-casino-black font-display tracking-widest
-                       text-sm py-4 rounded-lg transition-colors hover:bg-gold-light"
-          >
-            NEW GAME
-          </button>
-        </div>
-      </main>
-    </div>
+    <CompletePhase
+      group={group}
+      sessionPlayers={sessionPlayers}
+      onNewGame={() => {
+        setPhase("setup");
+        setSessionPlayers([]);
+      }}
+    />
   );
 }
